backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
pydantic>=2.5.3
pydantic-settings>=2.1.0
fastapi-mail>=1.4.1
python-dotenv>=1.0.0
aiosqlite
reportlab>=0.19.0
